requirements.txt
arrow==0.12.1
kafka-python==1.4.4
pathlib==1.0.1
psycopg2-binary==2.7.6.1
pymongo==3.6.1
python-dateutil==2.7.0
pytz==2018.3
six==1.12.0
SQLAlchemy==1.2.15
tornado==5.0
urllib3==1.22
wget==3.2
